requirements.txt
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==6.2
flake8==4.0.1
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.7.1
matplotlip==0.2
mccabe==0.6.1
mypy==0.920
mypy-extensions==0.4.3
numpy==1.21.4
packaging==21.3
pandas==1.3.5
platformdirs==2.4.0
plotly==5.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
tenacity==8.0.1
toml==0.10.2
tomli==2.0.0
typed-ast==1.5.1
typing_extensions==4.0.1
wrapt==1.13.3
zipp==3.6.0
